# swagger/spec.json
{"openapi":"3.0.0","info":{"title":"API SmartCity","version":"1.0.0"},"paths":{"/category/":{"get":{"tags":["Category"],"parameters":[{"name":"rowLimit","description":"Number of rows returned","in":"query","required":false,"schema":{"type":"integer"}},{"name":"offset","description":"Offset from the first line","in":"query","required":false,"schema":{"type":"integer"}},{"name":"searchElem","description":"Research element","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"$ref":"#/components/responses/CategoriesFound"},"400":{"$ref":"#/components/responses/CategoryBadParams"},"404":{"description":"No category found"},"500":{"description":"Server error"}}},"patch":{"tags":["Category"],"security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/CategoryToUpdate"},"responses":{"204":{"$ref":"#/components/responses/CategoryUpdated"},"400":{"description":"JWT not valid or JSON body not correct","content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/responses/ErrorJWT"},{"$ref":"#/components/responses/CategoryBadJSONBody"}]}}}},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeAdmin"},"404":{"description":"No category found"},"409":{"$ref":"#/components/responses/CategoryAlreadyExist"},"500":{"description":"Server error"}}},"post":{"tags":["Category"],"security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/CategoryToAdd"},"responses":{"201":{"$ref":"#/components/responses/CategoryAdded"},"400":{"description":"JWT not valid or JSON body not correct","content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/responses/ErrorJWT"},{"$ref":"#/components/responses/CategoryBadJSONBody"}]}}}},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeAdmin"},"409":{"$ref":"#/components/responses/CategoryAlreadyExist"},"500":{"description":"Server error"}}}},"/category/count":{"get":{"tags":["Category"],"security":[{"bearerAuth":[]}],"parameters":[{"name":"searchElem","description":"Research element","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"$ref":"#/components/responses/CategoryNumberFound"},"400":{"$ref":"#/components/responses/ErrorJWT"},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeAdmin"},"404":{"description":"Not found"},"500":{"description":"Server error"}}}},"/category/{id}":{"get":{"tags":["Category"],"security":[{"bearerAuth":[]}],"parameters":[{"name":"id","description":"category ID","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"$ref":"#/components/responses/CategoryFound"},"400":{"$ref":"#/components/responses/ErrorJWT"},"401":{"$ref":"#/components/responses/MissingJWT"},"404":{"description":"Category not found"},"500":{"description":"Server error"}}}},"/category":{"delete":{"tags":["Category"],"security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/CategoryToDelete"},"responses":{"204":{"$ref":"#/components/responses/CategoryDeleted"},"400":{"$ref":"#/components/responses/ErrorJWT"},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeAdmin"},"404":{"description":"No category found"},"500":{"description":"Erreur serveur"}}}},"/meal/":{"get":{"tags":["Meal"],"parameters":[{"name":"rowLimit","description":"Number of rows returned","in":"query","required":false,"schema":{"type":"integer"}},{"name":"offset","description":"Offset from the first line","in":"query","required":false,"schema":{"type":"integer"}},{"name":"searchElem","description":"Research element","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"$ref":"#/components/responses/MealsFound"},"400":{"$ref":"#/components/responses/MealBadParams"},"404":{"description":"No meal found"},"500":{"description":"Server error"}}},"patch":{"tags":["Meal"],"security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/MealToUpdate"},"responses":{"204":{"$ref":"#/components/responses/MealUpdated"},"400":{"description":"JWT not valid or JSON body not correct","content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/responses/ErrorJWT"},{"$ref":"#/components/responses/MealBadJSONBody"}]}}}},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeAdmin"},"404":{"description":"No Meal found"},"500":{"description":"Server error"}}},"post":{"tags":["Meal"],"security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/MealToAdd"},"responses":{"201":{"$ref":"#/components/responses/MealAdded"},"400":{"description":"JWT not valid or JSON body not correct","content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/responses/ErrorJWT"},{"$ref":"#/components/responses/MealBadJSONBody"}]}}}},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeAuthorizedOnRoute"},"500":{"description":"Server error"}}}},"/meal/count":{"get":{"tags":["Meal"],"security":[{"bearerAuth":[]}],"parameters":[{"name":"searchElem","description":"Research element","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"$ref":"#/components/responses/MealNumberFound"},"400":{"$ref":"#/components/responses/ErrorJWT"},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeAdmin"},"404":{"description":"Not found"},"500":{"description":"Server error"}}}},"/meal/{id}":{"get":{"tags":["Meal"],"security":[{"bearerAuth":[]}],"parameters":[{"name":"id","description":"Meal ID","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"$ref":"#/components/responses/MealFound"},"400":{"$ref":"#/components/responses/ErrorJWT"},"401":{"$ref":"#/components/responses/MissingJWT"},"404":{"description":"Meal not found"},"500":{"description":"Server error"}}}},"/meal":{"delete":{"tags":["Meal"],"security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/MealToDelete"},"responses":{"204":{"$ref":"#/components/responses/MealDeleted"},"400":{"$ref":"#/components/responses/ErrorJWT"},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeAdmin"},"404":{"description":"No Meal found"},"500":{"description":"Erreur serveur"}}}}},"components":{"schemas":{"Category":{"type":"object","properties":{"id":{"type":"integer"},"name":{"type":"string","description":"Category name"}}},"Meal":{"type":"object","properties":{"id":{"type":"integer"},"name":{"type":"string","description":"Meal name"},"description":{"type":"string","description":"Meal description"},"portion_number":{"type":"integer","description":"Number of portion of the meal"},"publication_date":{"type":"string","description":"When the meal was published"},"user_fk":{"type":"integer","description":"User ID of who created the meal"},"category_fk":{"type":"integer","description":"Category ID to which the meal belongs"},"order_fk":{"type":"integer","description":"Order ID of the related order (null if there is none)"},"image":{"type":"string","description":"Image path + name"}}}},"responses":{"CategoryBadJSONBody":{"description":"The JSON body is not correct"},"CategoryBadParams":{"description":"The URL parameters are not valid"},"CategoryAdded":{"description":"The category has been added"},"CategoryAlreadyExist":{"description":"Category name already taken"},"CategoryUpdated":{"description":"The category has been updated"},"CategoriesFound":{"description":"Return an array of categories","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Category"}}}}},"CategoryNumberFound":{"description":"Return the number of categories","content":{"application/json":{"schema":{"type":"integer"}}}},"CategoryFound":{"description":"Return a category","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Category"}}}},"CategoryDeleted":{"description":"The category has been deleted"},"MealBadJSONBody":{"description":"The JSON body is not correct"},"MealBadParams":{"description":"The URL parameters are not valid"},"MealAdded":{"description":"The meal has been added"},"MealUpdated":{"description":"The meal has been updated"},"MealsFound":{"description":"Return an array of meals","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Meal"}}}}},"MealNumberFound":{"description":"Return the number of meals","content":{"application/json":{"schema":{"type":"integer"}}}},"MealFound":{"description":"Return a meal","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Meal"}}}},"MealDeleted":{"description":"The meal has been deleted"},"mustBeAdmin":{"description":"This request can only be done by an administrator"},"mustBeAuthorizedOnRoute":{"description":"This request can only be done by an administrator or an authorized user"},"ErrorJWT":{"description":"JWT not valid"},"MissingJWT":{"description":"JWT not present"}},"requestBodies":{"CategoryToAdd":{"content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string"}},"required":["name"]}}}},"CategoryToUpdate":{"content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer","description":"Category ID"},"name":{"type":"string","description":"Category name"}},"required":["id","name"]}}}},"CategoryToDelete":{"content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer","description":"Category ID"}},"required":["id"]}}}},"MealToAdd":{"content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string","description":"Meal name"},"description":{"type":"string","description":"Meal description"},"portion_number":{"type":"integer","description":"Number of portion of the meal"},"publication_date":{"type":"string","description":"When the meal was published (omit for today date)"},"user_fk":{"type":"integer","description":"User ID of who created the meal"},"category_fk":{"type":"integer","description":"Category ID to which the meal belongs"},"order_fk":{"type":"integer","description":"Order ID of the related order (omit if there is none)"},"image":{"type":"object","description":"Image bytes"}},"required":["id","name","description","portion_number","category_fk","image"]}}}},"MealToUpdate":{"content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer"},"name":{"type":"string","description":"Meal name"},"description":{"type":"string","description":"Meal description"},"portion_number":{"type":"integer","description":"Number of portion of the meal"},"publication_date":{"type":"string","description":"When the meal was published"},"user_fk":{"type":"integer","description":"User ID of who created the meal"},"category_fk":{"type":"integer","description":"Category ID to which the meal belongs"},"order_fk":{"type":"integer","description":"Order ID of the related order (omit if there is none)"},"image":{"type":"object","description":"Image bytes (omit if you don't want to update the image)"}},"required":["id","name","description","portion_number","publication_date","user_fk","category_fk"]}}}},"MealToDelete":{"content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer","description":"Meal ID"}}}}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}},"tags":[]}